Filter immunisation coverage records

Select records from 'ic' objects based on region group, country, time, or vaccine.

ic_filter(X, region, country, time, vaccine)

# S3 method for ic.df
ic_filter(X, region, country, time, vaccine)

# S3 method for icfit
ic_filter(X, country, time, vaccine)

# S3 method for iclist
ic_filter(X, region, country, time, vaccine)

Arguments

X
object of type ic.df, icfit, or iclist

region
Character vector of regional groups to select. Not valid for icfit.

country
Character vector of country identifiers to select

time
Vector of time points to select

vaccine
Character vector of vaccine records to select, given by abbreviations (e.g. 'DTP1')

Value

A new object with type matching X with the selected records.

Details

For icfit and iclist objects, filtering will only affect the posterior estimates and prediction (if present). When a selection parameter is omitted, then all possible values will be selected. Selection parameters which are not present in X will be ignored.