Select survey records from an ic_survey dataset

Reduce immunisation survey coverage records based on set criteria of sample size and/or priority types of reporting.

survey_reduce(X, minSample = 300, priority = c("card or history", "card"))

Arguments

X
Object of type ic.df created with ic_survey.

minSample
Minimum number of samples in the survey.

priority
Character vector of the priority of records when there are duplicates records for a given vaccine, survey, and year. Priority is established by the 'evidence' field of many household surveys. The highest priority record is given first in this argument.

Value

An ic.df object with possibly fewer records, based on the selection criteria.

See also