Widely applicable information criterion (WAIC)

Implementation of waic() methods from loo to compute WAIC from the pointwise log-likelihood.

# S3 method for icfit
waic(X, ...)

# S3 method for iclist
waic(X, ...)

Arguments

X
Object of type icfit or iclist.

...
Additional arguments. See loo::waic().

Value

A named list (of class c("waic", "loo")) with components:

estimates

A matrix with two columns ("Estimate", "SE") and three rows ("elpd_waic", "p_waic", "waic"). This contains point estimates and standard errors of the expected log pointwise predictive density (elpd_waic), the effective number of parameters (p_waic) and the information criterion waic (which is just -2 * elpd_waic, i.e., converted to deviance scale).

pointwise

A matrix with three columns (and number of rows equal to the number of observations) containing the pointwise contributions of each of the above measures (elpd_waic, p_waic, waic).

Details

For more details see loo. The developers of stan recommend LOO-CV using PSIS (as implemented by the loo::loo() function) because PSIS provides useful diagnostics as well as effective sample size and Monte Carlo estimates.

References

Watanabe, S. (2010). Asymptotic equivalence of Bayes cross validation and widely application information criterion in singular learning theory. Journal of Machine Learning Research 11, 3571-3594.

See also